--- amazonCrawlers.py
@@ -1,350 +1,0 @@
-from selenium.common.exceptions import TimeoutException
-from selenium.webdriver.common.by import By
-from selenium.webdriver.support import expected_conditions as EC
-from selenium.webdriver.support.ui import WebDriverWait
-from selenium import webdriver
-from bs4 import BeautifulSoup
-from selenium.webdriver.firefox.options import Options
-from selenium.webdriver import Firefox
-from selenium.webdriver.common.keys import Keys
-import lxml.html
-#import pymongo
-import re
-from openpyxl import Workbook
-from openpyxl import load_workbook
-from openpyxl.compat import range
-from openpyxl.utils import get_column_letter
-import string
-#Time
-from datetime import datetime, date, time
-import time
-# Use to make a sound
-import os
-targetProductNameMatching = 'Maevis Bed Waterproof Mattress'
-
-#Headless Chrome
-options = webdriver.ChromeOptions()
-"""options.add_argument('headless')
-# Load no image makes it run faster
-prefs = {"profile.managed_default_content_settings.images":2}
-options.add_experimental_option("prefs",prefs)"""
-options.add_argument('window-size=1200x600')
-browser = webdriver.Chrome(chrome_options=options)
-wait = WebDriverWait(browser, 10)
-#Headless Firefox
-"""options = Options()
-options.add_argument('-headless')
-browser = Firefox(executable_path='geckodriver', firefox_options=options)
-browser.set_window_size(1200, 600)
-wait = WebDriverWait(browser, 10)"""
-
-#Excel part 
-#Global variable
-wb = Workbook()
-products = []
-
-def search(keyword,pageNumber):
-    print('正在搜索')
-    # Start
-    os.system('say "Your program is start it fkuc!"')
-    try:
-        #美亚
-        browser.get('https://www.amazon.com/')
-        input = wait.until(
-            EC.presence_of_element_located((By.CSS_SELECTOR, '#twotabsearchtextbox')))
-        submit = wait.until(
-            EC.element_to_be_clickable((By.CSS_SELECTOR, '.nav-search-submit > input:nth-child(2)')))
-        input.send_keys(keyword)
-        submit.click()
-        get_products(keyword,pageNumber)
-        # Test to compare pages
-        #time.sleep(100)
-    except TimeoutException:
-        return search(keyword,pageNumber)
-
-
-def next_page(keyword,pageNumber):
-    print('正在翻页', pageNumber)
-    try:
-        wait.until(EC.text_to_be_present_in_element(
-            (By.CSS_SELECTOR, '#pagnNextString'), 'Next Page'))
-        submit = wait.until(EC.element_to_be_clickable((By.CSS_SELECTOR, '#pagnNextString')))
-        submit.click()
-        wait.until(EC.text_to_be_present_in_element(
-            (By.CSS_SELECTOR, '.pagnCur'), str(pageNumber)))
-        get_products_title_index(keyword,pageNumber)
-        """ 位置测试 （测试环境- Google Chrome）
-        # 测试发现 除了广告位会变动外
-        # 自然位的不会变动 一切匹配
-        # 运行正常"""
-    except TimeoutException:
-        next_page(keyword,pageNumber)
-
-
-def get_products_title_index(keyword,pageNumber):
-    try:
-        # 使用默认的排列
-        wait.until(EC.presence_of_element_located((By.CSS_SELECTOR, '#s-results-list-atf')))
-        html = browser.page_source
-        soup = BeautifulSoup(html, 'lxml')
-        #result_0是第一个
-        content = soup.find_all(attrs={"id": re.compile(r'result_\d+')})
-        print("how many result were found:",len(content))
-        # 如果行数超过13行 发出警报
-        if len(content) > 45:
-            print('More than 15 rows.')
-            os.system('say "It is more than 15 rows, you should check if it is a BUG!"')
-        #如果有搜索结果
-        if len(content)!=0:
-            #这里除了问题 url和content数量不一致 导致url少的先结束for loop
-            for index,item in enumerate(content):
-                product = {
-                    #是那种非产品的缺少s-access-title的，就默认给个title
-                    'title': item.find(class_='s-access-title').get_text() if item.find(class_='s-access-title') else "Amazon recommendation", 
-                    'index': index,#在一页里的顺位序号，每一页都会变
-                    #'rank': getRank(pageNumber,index),#就算有那种AD也是准的，不影响
-                }
-                #Generate Rank attr for product
-                turnProductIndexToRank(product,pageNumber)
-                products.append(product)
-                #print(index+1," product is processed.")
-        else:
-            print("No products were found!")
-        #print("how many prodcuts:",len(products))"""
-    except Exception as err:
-        print(err)
-
-#BUG-因为selenium和Chrome，firefox等显示的页面不一样，selenium上展示的产品会偏多
-#所以无法判断位置
-#https://www.quora.com/How-come-the-page-source-from-the-webpage-is-different-when-I-view-it-on-Chrome-compared-to-when-Selenium-parses-it
-def turnProductIndexToRank(product,pageNumber):
-    #Make the soup
-    html = browser.page_source
-    soup = BeautifulSoup(html, 'lxml')
-
-    # 如果是九宫格和四宫格都有 默认的展示方式就可以了
-    if soup.find('div',class_="s-grid-layout-picker"):
-        # 选9宫格模式
-        # 同时有9宫格和4宫格 
-        # 都是3列 
-        if soup.find('div',class_='s-image-layout-picker'):
-            productIndex = product['index']
-            if productIndex <= 3:
-                product['rank'] = str(pageNumber)+","+"1"+","+str(productIndex)
-            # 3的倍数    
-            elif productIndex%3 ==0:
-                product['rank'] = str(pageNumber)+","+str(productIndex/3)+","+"3"
-            else:
-                product['rank'] = str(pageNumber)+","+str(productIndex//3 + 1)+","+str(productIndex%3)
-    # 剩下的就是列模式了
-    # 1_列模式可翻页的那种模式
-    #如:https://www.amazon.com/s/ref=nb_sb_noss_2?url=search-alias%3Daps&field-keywords=tv&rh=i%3Aaps%2Ck%3Atv&ajr=0
-    # 同时有四格和列
-    elif soup.find('div',class_='s-list-layout-picker'):
-        if soup.find('div',class_='s-image-layout-picker'):
-            product['rank'] =  str(pageNumber)+','+str(productIndex)
-    # 列
-    # https://www.amazon.com/s?field-keywords=sleeping+bag
-    elif not soup.find('div',class_='s-list-layout-picker'):
-        if not soup.find('div',class_='s-image-layout-picker'):
-            if not soup.find('div',class_="s-grid-layout-picker"):
-                if soup.find('span',id='a-autoid-0-announce').get_text() != 'See more':
-                    os.system('say "Maybe it is a coloum mode, please check"')
-                    print("Check please!!!")
-                    product['rank'] =  str(pageNumber)+','+str(productIndex)
-    else:
-        #2_see more的那种像厕纸一样的中间分页的那种没有翻页的那种列模式
-        #如：https://www.amazon.com/gp/vs/buying-guide/sleeping-bag/459108?ie=UTF8&field-keywords=sleeping%20bag&ref_=nb_sb_ss_ime_c_1_9&url=search-alias%3Daps
-        #TODO: 什么时候解决下
-        #soup.find('span',id='a-autoid-0-announce').get_text() == 'See more':
-        #print("See more mode")
-        #product['rank'] = "See more mode"
-        #Log到Excel的rank那里表示遇到了这种情况
-        print("Not the normal 3 modes")
-        product['rank'] = "Other mode"
-
-
-#Save Rank to Excel
-def saveRankToExcel(products,pageNumber,keyword):
-    try:
-        for product in products:
-            productTitle = product['title']
-            productRank = product['rank']
-            wb[keyword].append([productTitle,productRank])
-        wb.save("sample.xlsx")
-        print('Saved')
-    except Exception as err:
-        print('Save Rank failed:', err)   
-        wb.save("sample.xlsx")  
-
-#获取并储存第一个广告和自然搜索的位置
-#BUG-判断是否是自己的产品时出现了失误
-def saveFirstAD_nonAD_rankToExcelIn1stSheet(products,whichKindOfProduct,keyword,keywordIndex):
-    try:
-        if len(products) == 0:
-            print('Nothing is catched.')
-            return
-        #TODO-优化，不要遍历所有的 找到两个就停止
-        """#一个广告和一个自然
-        twoProducts = []
-        #如果找到了两个就停止loop
-        stopSign = len(twoProducts)
-        if stopSign == 2:
-                print("Two were found,stop searching.")
-                break"""
-        #注意这个部分
-        #亚马逊后台SKU的标题改变的话 这部分字典也要及时修改
-        #防水床笠(fscl首字母缩写)部分
-        fscl = {
-            'Maevis Bed Waterproof Mattress Protector Cover Pad Fitted 18 Inches Deep Pocket Premium Washable Vinyl Free - Twin XL':'TXL',
-            'Waterproof Mattress Cover Protector Pad with 18 Inches Deep Pocket for Full Bed by Maevis,Full Size':'F',
-            'Maevis Bed Waterproof Mattress Protector Cover Pad Fitted 18 Inches Deep Pocket Premium Washable Vinyl Free - Queen':'Q',
-            'Maevis Bed Waterproof Mattress Protector Cover Pad Fitted 18 Inches Deep Pocket Premium Washable Vinyl Free - King':'K',
-            'Maevis Bed Waterproof Mattress Protector Cover Pad Fitted 18 Inches Deep Pocket Premium Washable Vinyl Free - California King':'CK'
-        }
-        #夹棉床笠部分
-        jmcl = {
-            'Maevis Mattress Pad Cover 100% 300TC Cotton with 8-21 Inch Deep Pocket White Overfilled Bed Mattress Topper (Down Alternative, Twin)':'T',
-            'Maevis Mattress Pad Cover 100% 300TC Cotton with 8-21 Inch Deep Pocket White Overfilled Bed Mattress Topper (Down Alternative, Twin XL)':'TXL',
-            'Maevis Mattress Pad Cover 100% 300TC Cotton with 8-21 Inch Deep Pocket White Overfilled Bed Mattress Topper (Down Alternative, Full)':'F',
-            'Maevis Mattress Pad Cover 100% 300TC Cotton with 8-21 Inch Deep Pocket White Overfilled Bed Mattress Topper (Down Alternative, Queen)':'Q',
-            'Maevis Mattress Pad Cover 100% 300TC Cotton with 8-21 Inch Deep Pocket White Overfilled Bed Mattress Topper (Down Alternative, King)':'K',
-            'Maevis Mattress Pad Cover 100% 300TC Cotton with 8-21 Inch Deep Pocket White Overfilled Bed Mattress Topper (Down Alternative, California King)':'CK'
-        }
-        """#26
-        numberAlphabetMapping = {
-            '1':'A','2':'B','3':'C','4':'D','5':'E',
-            '6':'F','7':'G','8':'H','9':'I','10':'J',
-            '11':'K','12':'L','13':'M','14':'N','15':'O',
-            '16':'P','17':'Q','18':'R','19':'S','20':'T',
-            '21':'U','22':'V','23':'W','24':'X','25':'Y','26':'Z'
-        } """
-        ad = []
-        nonAd = []
-        targetAdRank = ''
-        targetAdSize = ''
-        targetNonAdRank = ''
-        targetNonAdSize = ''
-        productType = {}
-        unifiedRankAndSize = ''
-        #先找到自己的产品
-        #寻找第一个广告和自然搜索的位置
-        if whichKindOfProduct == 'fscl':
-            productType = fscl
-        if whichKindOfProduct == 'jmcl':
-            productType = jmcl
-        print('product type:',productType)
-        for product in products:
-            #Identify and sort our products
-            #Identify
-            if product['title'] in list(productType):
-                print("product title:",product['title'])
-                #Sort the products to ad and non-ad
-                if "Sponsored" in product['title']:
-                    ad.append(product)
-                else:
-                    nonAd.append(product)
-       
-        #Process data
-        if len(ad) != 0:
-            targetAdRank = str(ad[0]['rank'])
-            targetAdSize = productType[ad[0]['title']] + '广告'
-        if len(nonAd) != 0:
-            targetNonAdRank = str(nonAd[0]['rank'])
-            targetNonAdSize = productType[nonAd[0]['title']] + '自然'
-  
-        #print("AD",len(ad))
-        #print('NonAd',len(nonAd))
-        #Unify
-        unifiedRankAndSize = targetAdRank+'('+targetAdSize+')'+'/'+targetNonAdRank+'('+targetNonAdSize+')'
-        print("unifyied Rank",unifiedRankAndSize)
-        if unifiedRankAndSize == "()/()":
-            unifiedRankAndSize = '大于8页'
-        #Because the first cell is occupied by today's date
-        rankRow = 2
-        rankColoum = keywordIndex+1
-        keywordRow = 1
-        keywordColoum = rankColoum
-        #keyword title cell
-        wb['Sheet'].cell(keywordRow,keywordColoum,keyword)
-        #keyword rank cell
-        wb['Sheet'].cell(rankRow,rankColoum,unifiedRankAndSize)
-        wb.save("sample.xlsx")
-        return unifiedRankAndSize
-    except Exception as err:
-        print('Save to 1st sheet failed:', err)   
-        wb.save("sample.xlsx") 
-    
-# Title with Rank
-def main():
-    try:
-        startTime = datetime.now()
-        print("Start at:",startTime)
-
-        global products
-        # keywords to search
-        #夹棉床笠
-        keywords = ['mattress protector','queen mattress pad','mattress topper','queen mattress topper','twin mattress pad','king mattress pad','mattress cover','mattress pad cover']
-        #keywords = ['sheets']
-        #keywords = ['mattress protector','queen mattress pad']
-        #There are 'fscl' 防水床笠 and 'jmcl' 夹棉床笠 这两个 type
-        #whichKindOfProduct = 'fscl'
-        #keywords = ['mattress protector']
-        #whichKindOfProduct = 'jmcl' 
-        #wb['Sheet'].cell(2,1,startTime)
-        for keyword in keywords:
-            #global products
-            # one keyword per sheet
-            #ws = wb.create_sheet(title=keyword)
-            #ws.append(["Product Name", "Star Rank"])
-            #Reset pageNumber when keyword changed
-            pageNumber = 1
-            search(keyword,pageNumber)
-            #Display only the first N pages
-            for pageNumber in range(2, 8):
-                next_page(keyword,pageNumber)
-            #Done getting data
-            #Persist data to excel
-            #saveRankToExcel(products,pageNumber,keyword)
-            #Process all prodcuts obtained
-            #keywordIndex = sheetx
-            #saveFirstAD_nonAD_rankToExcelIn1stSheet(products,whichKindOfProduct,keyword,keywordIndex)
-            #重置products 如果关键词多的话 需要重置
-            #products = []
-            #print("if its 0 then it's reset",len(products))
-        endTime = datetime.now()
-        print("Ends at:",endTime)
-    except Exception as err:
-        print('出错啦', err)
-        endTime = datetime.now()
-        print("Ends at:",endTime)
-        #wb.save("sample.xlsx")
-    finally:
-        #browser.quit()
-        pass
-#BUG-有的界面没有那个九宫格显示模式，怎么强制切换。
-#TODO:添加一个处理总时
-#TODO:保存一个条目时保存一下
-#TODO:'NoneType' object has no attribute 'get_text' 处理下 排查下
-if __name__ == '__main__':
-    main()
-
-
-#BUG-出错啦 Message: Timeout loading page after 300000ms
-#BUG-Fixed不能在这里退出浏览器 不然不能搜其他的产品连接了
-#browser.quit()
-#注意Openpyxl添加行时.append([])要添加一个list
-#.append()是添加一个单元格
-
-#TODO:sleeping bag这种Rank如何计算
-#Not the normal 3 modes
-#Save Rank failed: local variable 'product' referenced before assignment
-#增加进度条 不然不知道是不是卡住了
-#创建Github分支不要直接在Main分支操作
-
-#非正常的3行模式product的提取方式也不一样
-
-#BUG-EXCEL结果和网页不一样排序
-#TODO：需要翻8页很浪费 处理下如何在找到2个最靠前的自然位后停止翻页
-
-#
